Reimplement the filetree as a C++ class
Re-phrasing this app as Direct2D/C++ seems to be giving crazy good development/performance results.

Runtime RAM in release is maxxing at about 20MB and CPU usage at about 5% according to task manager (and 0% GPU usage, of course).

Rendering would easily become one of the greatest performance tanks to the application, so we'll probably have to be smart with using a retained mode UI and Invalidate(Rect) to keep CPU usage to a minimum.